CRUD completed

+++ CRUDlibrary/Entity.cs
@@ -4,12 +4,12 @@ namespace CRUDlibrary
 {
     public class Busdetails
     {
-        public long BusID { get; set; }
+        public int BusID { get; set; }
         public string BusName { get; set; }
         public long DriverMobilenumber { get; set; }
         public string StartPoint { get; set; }
         public string Destination { get; set; }
-        public long Fair { get; set; }
-        public long NoofPassenger { get; set; }
+        public int Fair { get; set; }
+        public int NoofPassenger { get; set; }
     }
 }
